etc/face-api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
numpy==1.24.3
Pillow==10.4.0
pydantic==2.9.2
gunicorn==21.2.0
